Moved roll action to it's own file - changed css for displaying dice
